[dataset] Move generic Dataset class to local __init__.py module
